new filter

+++ filter.py
@@ -5,11 +5,20 @@ parser=OptionParser()
 parser.add_option("-f", action = "store", type = "string", dest = "filename")
 parser.add_option("-y", action = "store", type = "int", dest = "year")
 (options,args) = parser.parse_args()
+wd = os.getcwd()
+FN = options.filename
+FN = "evolving/"+ str(options.year)+"/"+FN 
+if os.path.exists(wd+'/evolving'):
+	if os.path.exists(wd+'/evolving'+'/'+str(options.year)):
+		wf = open(FN , 'w')
+	else:
+		os.makedirs(wd + "/evolving/" + str(options.year))
+		wf = open(FN , 'w')
+else:
+	os.makedirs(wd + "/evolving/" + str(options.year))
+	wf = open(FN , 'w')
 
-FN = options.filename.split('.')
-FN[0] = FN[0] + "_" + str(options.year)
-FN = '.'.join(FN)
-wf = open(FN , 'w')
+
 with open(options.filename) as f:
 	header = f.readline()
 	wf.write(header)
@@ -19,7 +28,7 @@ with open(options.filename) as f:
 			break
 		else:
 			l = l.split(',')
-			if int(l[6]) <= options.year:
+			if int(l[6]) >= options.year:
 				l = ','.join(l)
 				wf.write(l)
 f.close()
